Refactor Climate Advisor integration for Stationary Energy drafts
- Reorganized API calls to utilize shared client options for OpenRouter, enhancing maintainability.
- Introduced new prompt for generating stationary energy draft proposals, improving draft creation capabilities.
- Streamlined message handling in chat routes, ensuring consistent payload structures.
- Updated tests to reflect changes in API usage and client configuration, ensuring robust coverage.
- Enhanced documentation for clarity on new features and usage patterns.
