+++ climate-advisor/service/pytest.ini
@@ -18,6 +18,9 @@ markers =
 
 # Asyncio configuration
 asyncio_mode = auto
+pythonpath =
+    .
+    ..
 
 # Output settings
 addopts =
