+++ climate-advisor/service/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Climate Advisor test modules and runnable test utilities."""
