+++ climate-advisor/service/tests/test_cc_inventory_tool.py
@@ -1,15 +1,7 @@
 from __future__ import annotations
 
 import json
-import sys
-from pathlib import Path
 import unittest
-
-PROJECT_ROOT = Path(__file__).resolve().parents[2]
-for extra_path in (PROJECT_ROOT, PROJECT_ROOT / "service"):
-    path_str = str(extra_path)
-    if path_str not in sys.path:
-        sys.path.insert(0, path_str)
 
 from app.services.citycatalyst_client import (
     CityCatalystClientError,
