+++ climate-advisor/service/tests/test_e2e_conversation.py
@@ -12,21 +12,13 @@ Tests cover:
 
 from __future__ import annotations
 
-import sys
 import json
-from pathlib import Path
 from typing import AsyncIterator
 from unittest.mock import AsyncMock, patch, MagicMock
 import unittest
 
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 from sqlalchemy.pool import StaticPool
-
-PROJECT_ROOT = Path(__file__).resolve().parents[2]
-for extra_path in (PROJECT_ROOT, PROJECT_ROOT / "service"):
-    path_str = str(extra_path)
-    if path_str not in sys.path:
-        sys.path.insert(0, path_str)
 
 from app.main import get_app
 from app.db import Base
