+++ climate-advisor/service/tests/test_prompt_budget.py
@@ -1,14 +1,5 @@
 from __future__ import annotations
 
-import sys
-from pathlib import Path
-
-PROJECT_ROOT = Path(__file__).resolve().parents[2]
-for extra_path in (PROJECT_ROOT, PROJECT_ROOT / "service"):
-    path_str = str(extra_path)
-    if path_str not in sys.path:
-        sys.path.insert(0, path_str)
-
 from app.config.settings import _load_llm_config
 from app.utils.prompt_budget import (
     StationaryEnergyPromptBudget,
