+++ climate-advisor/service/tests/test_rag_tool.py
@@ -5,19 +5,10 @@ interface retrieves full chunk content for retrieval-augmented generation.
 
 from __future__ import annotations
 
-import sys
-from pathlib import Path
 from types import SimpleNamespace
 from typing import Any, Dict, List, Optional
 import unittest
 from unittest.mock import AsyncMock
-
-
-PROJECT_ROOT = Path(__file__).resolve().parents[2]
-for extra_path in (PROJECT_ROOT, PROJECT_ROOT / "service"):
-    path_str = str(extra_path)
-    if path_str not in sys.path:
-        sys.path.insert(0, path_str)
 
 from app.tools.climate_vector_tool import ClimateToolResult, ClimateVectorSearchTool
 from app.config.settings import _load_llm_config
