+++ climate-advisor/service/tests/test_vector_db_insertion.py
@@ -11,7 +11,6 @@ from __future__ import annotations
 
 import asyncio
 import os
-import sys
 import unittest
 from pathlib import Path
 from typing import List
@@ -25,13 +24,6 @@ from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_asyn
 env_path = Path(__file__).resolve().parents[2] / '.env'
 load_dotenv(env_path)
 
-# Ensure project root is importable
-PROJECT_ROOT = Path(__file__).resolve().parents[2]
-if str(PROJECT_ROOT) not in sys.path:
-    sys.path.insert(0, str(PROJECT_ROOT))
-
-# Import the correct model (service app version)
-sys.path.insert(0, str(PROJECT_ROOT / "service"))
 from app.models.db.document_embedding import DocumentEmbedding
